Fix mypy issues; rename util modules
