+++ src/backend/aspen/api/utils/__init__.py
@@ -1,9 +1,7 @@
 from aspen.api.utils.authz import authz_samples_cansee  # noqa: F401
-from aspen.api.utils.format_date import format_date, format_datetime  # noqa: F401
-from aspen.api.utils.format_sample import (  # noqa: F401
+from aspen.api.utils.date import format_date, format_datetime  # noqa: F401
+from aspen.api.utils.gisaid import get_matching_gisaid_ids  # noqa: F401
+from aspen.api.utils.sample import (  # noqa: F401
     determine_gisaid_status,
     format_sample_lineage,
-)
-from aspen.api.utils.get_matching_gisaid_ids import (  # noqa: F401
-    get_matching_gisaid_ids,
 )
